move MinAggregator and MaxAggregator to functions-aggregate-common
