+++ datafusion/datasource-parquet/Cargo.toml
@@ -40,7 +40,7 @@ datafusion-common-runtime = { workspace = true }
 datafusion-datasource = { workspace = true, features = ["parquet"] }
 datafusion-execution = { workspace = true }
 datafusion-expr = { workspace = true }
-datafusion-functions-aggregate = { workspace = true }
+datafusion-functions-aggregate-common = { workspace = true }
 datafusion-physical-expr = { workspace = true }
 datafusion-physical-expr-adapter = { workspace = true }
 datafusion-physical-expr-common = { workspace = true }
